requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
Werkzeug==3.0.1

# ML/AI Dependencies (Optional - for enhanced chatbot)
scikit-learn==1.3.0
numpy==1.24.3
# spacy==3.6.0  # Uncomment for advanced entity extraction
# sentence-transformers==2.2.2  # Uncomment for semantic similarity
